Fix access violation crash from CPredictionCopy_TransferData signature
- Add null pointer check in CPredictionCopy::TransferData to prevent crash when signature not found
- Add error handling in EnginePrediction Start/End to skip prediction on signature failure
- Add MemAlloc interface null check to prevent memory allocation crashes
- Fix size_t to int conversion warning in SpectateAll.cpp (line 297)
- Fix unreferenced variable warning in IBaseClientDLL_DispatchUserMessage.cpp

Resolves crash 0xc0000005 with RAX=0x0 caused by calling invalid CPredictionCopy_TransferData signature introduced in commit ec3b802.

+++ Amalgam/src/Features/Misc/SpectateAll/SpectateAll.cpp
@@ -289,7 +289,7 @@ CTFPlayer* CSpectateAll::GetNextEnemyPlayer()
     if (Vars::Competitive::SpectateAll::PreferKillerSpectate.Value && m_pLastKiller && m_pLastKiller->IsAlive())
     {
         // Find our killer in the enemies list
-        for (size_t i = 0; i < enemies.size(); i++)
+        for (int i = 0; i < static_cast<int>(enemies.size()); i++)
         {
             auto pEnemy = enemies[i]->As<CTFPlayer>();
             if (pEnemy == m_pLastKiller)
